gui/cmake: Add CMake build.
Adds CMakeLists.txt for the Compressonator GUI and it's dependencies
_Libs/CMP_MeshOptimizer, _Libs/GPU_Decode, _Plugins/Common/gltf.

I didn't plan to recreate the CMake build in the local copy of the Draco
library. This build is using the Draco library as an external dependency
via pkgconfig, since Draco's Find CMake files are broken. They could be
usable on Windows though.

Draco is perfectly installable and provides CMake and Pkgconfig files.
As an example I packaged Draco for ArchLinux, which can be found here:

https://aur.archlinux.org/packages/draco-git/