cli/cmake: Use CMake target names in favor of paths.
In CMake the target names can be passed to the linker directly in favor
of building the paths by hand. This is not only nicer to look at, but
also revolved build errors in QtCreator for me, which apparently did
some path magic.

This will work on all platforms and also when libraries are not build
as `.so` files.